重新安排 RBAC 错误定义文件
合并所有 RBAC 方面的错误定义到一个文件 - 减少引用混乱
