+++ leaf/rbac/__init__.py
@@ -1,5 +1,6 @@
 """用户用户组及认证相关 - 实现基于角色的用户权限控制"""
 
 from . import jwt
+from . import error
 from . import model
 from . import functions
