+++ leaf/rbac/functions/__init__.py
@@ -1,6 +1,5 @@
 """CRUD 使用到的函数集合"""
 
-from . import error
 from . import auth
 from . import user
 from . import group
